Pushing EmbDistRetrieve baseline along with a) DEPENDENCIES b) INSTRUCTIONS
Note that we have instructions to download post-processed input dumps for now
These dumps were also used for the numbers in the original paper, and can be directly used for the various StyleTransfer approaches
to replicate those numbers.

We are still readying our post-processing code and will share the full code [converting the raw dataset jsons to these dumps] as well as the
post-processed dumps for v1.2 , mostly within the coming week.